pyproject.toml
[tool.poetry]
name = "kitsune-selenium-tests"
version = "0.1.0"
description = "\"Selenium tests for Mozilla Kitsune\""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "kitsune_selenium_tests"}]

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.86.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
selenium = "^4.8.3"
pytest-html = "^3.2.0"
pytest-xdist = "^3.2.1"
docker = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
